refactor : API 클라이언트 구조 개선
- TmdbApi 클래스를 범용적인 ApiClient로 리팩토링
- API 종속적인 메서드 별도 함수로 분리
