+++ src/api/tmdb/tmdbApiClient.ts
@@ -1,0 +1,72 @@
+import ApiClient from '../ApiClient';
+
+export interface APIResponse<T> {
+  page: number;
+  results: T[];
+  total_pages: number;
+  total_results: number;
+}
+
+export interface MovieResponse {
+  id: number;
+  title: string;
+  original_title: string;
+  poster_path: string | null;
+  backdrop_path: string | null;
+  overview: string;
+  release_date: string;
+  vote_average: number;
+  vote_count: number;
+  popularity: number;
+  adult: boolean;
+  genre_ids: number[];
+  video: boolean;
+  original_language: string;
+}
+
+export interface MovieDetailResponse {
+  id: number;
+  title: string;
+  original_title: string;
+  poster_path: string | null;
+  backdrop_path: string | null;
+  overview: string;
+  release_date: string;
+  vote_average: number;
+  vote_count: number;
+  popularity: number;
+  adult: boolean;
+  budget: number;
+  genres: { id: number; name: string }[];
+  homepage: string | null;
+  imdb_id: string | null;
+  original_language: string;
+  revenue: number;
+  runtime: number | null;
+  status: string;
+  tagline: string | null;
+  video: boolean;
+  belongs_to_collection: any | null;
+  production_companies: {
+    id: number;
+    logo_path: string | null;
+    name: string;
+    origin_country: string;
+  }[];
+  production_countries: {
+    iso_3166_1: string;
+    name: string;
+  }[];
+  spoken_languages: {
+    english_name: string;
+    iso_639_1: string;
+    name: string;
+  }[];
+}
+
+const tmdbApiClient = new ApiClient(
+  import.meta.env.VITE_API_TOKEN || '',
+  import.meta.env.VITE_BASE_URL || '',
+);
+
+export default tmdbApiClient;
